requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
Pillow==10.2.0
python-multipart==0.0.9
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5 